moving the remaining of utils out of core
