+++ test/integration/02-user-wallet/02-receive-onchain.spec.ts
@@ -1,6 +1,6 @@
 import { once } from "events"
 import { getOnChainAddressCreateAttemptLimits, getUserLimits } from "@config/app"
-import { sleep } from "@core/utils"
+import { sleep } from "@app/utils"
 import { getTitle } from "@services/notifications/payment"
 import { onchainTransactionEventHandler } from "@servers/trigger"
 import {
