+++ test/integration/core/lock.spec.ts
@@ -1,5 +1,5 @@
 import { redis } from "@services/redis"
-import { sleep } from "@core/utils"
+import { sleep } from "@app/utils"
 import { baseLogger } from "@services/logger"
 import { redlock, getResource, lockExtendOrThrow } from "@core/lock"
