+++ test/integration/servers/graphql-main-server/auth-requests.spec.ts
@@ -1,5 +1,5 @@
 import { JWT_SECRET, yamlConfig } from "@config/app"
-import { sleep } from "@core/utils"
+import { sleep } from "@app/utils"
 import { startApolloServerForCoreSchema } from "@servers/graphql-main-server"
 import { createTestClient } from "apollo-server-integration-testing"
 import { createHttpTerminator } from "http-terminator"
