+++ test/integration/services/lnd/utils.spec.ts
@@ -6,7 +6,7 @@ import {
 } from "@services/lnd/utils"
 import { baseLogger } from "@services/logger"
 import { ledger } from "@services/mongodb"
-import { sleep } from "@core/utils"
+import { sleep } from "@app/utils"
 import { DbMetadata, InvoiceUser } from "@services/mongoose/schema"
 import {
   cancelHodlInvoice,
